Mask out unused pins AFTER XORing
Mask out unused pins after XORing. This was not a problem when
PRESSED = 1 because the debug pins would be masked out (all 0s),
then xored with 0xFF (all 1s), then negated (all 0s) so it wouldn't
affect the outcome.  When PRESSED = 0, though, the xor step does
nothing, so when negated its read as a controller input.